% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Calculate CV}
\usage{
cv(x = NULL, na_rm = TRUE)
}
\arguments{
\item{x}{Vector of numbers}

\item{na_rm}{Whether \code{NA}s should be removed. Defaults to \code{TRUE}}
}
\value{
Coefficient of variation of the numbers provided to 'x'
}
\description{
Calculates the Coefficient of Variation ("CV") of a vector of numbers.
}
\examples{
# Calculate CV
cv(x = c(4, 5, 6), na_rm = TRUE)
}
